plugin/ftp/lib: fixed clang warning
git-svn-id: svn://ultimatepp.org/upp/trunk@15841 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/plugin/ftp/lib/ftplib.h
@@ -62,7 +62,7 @@ GLOBALREF void FtpInit(void);
 GLOBALREF char *FtpLastResponse(netbuf *nControl);
 GLOBALREF int FtpConnect(const char *host, netbuf **nControl, char perror[512],
 	FtpCallback idlecb, void *idledata, int idletime_msecs, int idletimeout_secs);
-GLOBALREF int FtpOptions(int opt, long val, netbuf *nControl);
+GLOBALREF int FtpOptions(int opt, uintptr_t val, netbuf *nControl);
 GLOBALREF int FtpLogin(const char *user, const char *pass, netbuf *nControl);
 GLOBALREF int FtpAccess(const char *path, int typ, int mode, netbuf *nControl,
     netbuf **nData);
